SCB-327 adjust health check package
