+++ metrics/metrics-core/src/test/java/org/apache/servicecomb/metrics/core/TestHealthCheckerPublisher.java
@@ -20,36 +20,83 @@ package org.apache.servicecomb.metrics.core;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 import org.apache.servicecomb.foundation.common.utils.JsonUtils;
-import org.apache.servicecomb.foundation.metrics.publish.HealthCheckResult;
+import org.apache.servicecomb.foundation.metrics.health.HealthCheckResult;
+import org.apache.servicecomb.foundation.metrics.health.HealthChecker;
+import org.apache.servicecomb.foundation.metrics.health.HealthCheckerManager;
+import org.apache.servicecomb.metrics.core.health.DefaultHealthCheckExtraData;
+import org.apache.servicecomb.metrics.core.health.DefaultMicroserviceHealthChecker;
 import org.apache.servicecomb.metrics.core.publish.DefaultHealthCheckerPublisher;
-import org.apache.servicecomb.metrics.core.publish.HealthCheckerManager;
+import org.apache.servicecomb.serviceregistry.RegistryUtils;
+import org.apache.servicecomb.serviceregistry.api.registry.Microservice;
+import org.apache.servicecomb.serviceregistry.api.registry.MicroserviceInstance;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
+import com.google.common.collect.Lists;
+
+import mockit.Expectations;
 
 public class TestHealthCheckerPublisher {
 
   @Test
-  public void testPublisher() throws JsonProcessingException {
+  public void testPublisher() throws IOException {
+    Microservice microservice = new Microservice();
+    microservice.setAppId("appId");
+    microservice.setServiceName("serviceName");
+    microservice.setVersion("0.0.1");
+
+    MicroserviceInstance microserviceInstance = new MicroserviceInstance();
+    microserviceInstance.setEndpoints(Lists.newArrayList("127.0.0.1", "192.168.0.100"));
+    microserviceInstance.setInstanceId("001");
+    microserviceInstance.setHostName("localhost");
+
+    new Expectations(RegistryUtils.class) {
+      {
+        RegistryUtils.getMicroservice();
+        result = microservice;
+      }
+    };
+
+    new Expectations(RegistryUtils.class) {
+      {
+        RegistryUtils.getMicroserviceInstance();
+        result = microserviceInstance;
+      }
+    };
+
     HealthCheckerManager manager = mock(HealthCheckerManager.class);
 
     Map<String, HealthCheckResult> results = new HashMap<>();
-    HealthCheckResult result = new HealthCheckResult(true, "ok", "extra");
-    results.put("default", result);
+    HealthChecker result = new DefaultMicroserviceHealthChecker();
+    results.put("default", result.check());
 
     when(manager.check()).thenReturn(results);
-    when(manager.check("default")).thenReturn(result);
+    when(manager.check("default")).thenReturn(result.check());
 
     DefaultHealthCheckerPublisher publisher = new DefaultHealthCheckerPublisher(manager);
     Map<String, HealthCheckResult> content = publisher.health();
-    Assert.assertEquals(JsonUtils.writeValueAsString(result), JsonUtils.writeValueAsString(content.get("default")));
 
-    HealthCheckResult checkResult = publisher.healthWithName("default");
-    Assert.assertEquals(JsonUtils.writeValueAsString(result), JsonUtils.writeValueAsString(checkResult));
+    DefaultHealthCheckExtraData data = JsonUtils.OBJ_MAPPER
+        .readValue(content.get("default").getExtraData(), DefaultHealthCheckExtraData.class);
+    Assert.assertEquals("appId", data.getAppId());
+    Assert.assertEquals("serviceName", data.getServiceName());
+    Assert.assertEquals("0.0.1", data.getServiceVersion());
+    Assert.assertEquals("001", data.getInstanceId());
+    Assert.assertEquals("localhost", data.getHostName());
+    Assert.assertEquals("127.0.0.1,192.168.0.100", data.getEndpoints());
+
+    data = JsonUtils.OBJ_MAPPER
+        .readValue(publisher.healthWithName("default").getExtraData(), DefaultHealthCheckExtraData.class);
+    Assert.assertEquals("appId", data.getAppId());
+    Assert.assertEquals("serviceName", data.getServiceName());
+    Assert.assertEquals("0.0.1", data.getServiceVersion());
+    Assert.assertEquals("001", data.getInstanceId());
+    Assert.assertEquals("localhost", data.getHostName());
+    Assert.assertEquals("127.0.0.1,192.168.0.100", data.getEndpoints());
   }
 }
